+++ .gitignore
@@ -1,2 +1,3 @@
 /target
 .env
+new_img.*
