+++ src/services/mod.rs
@@ -1,0 +1,1 @@
+pub mod img_upload;
